requirements.txt
django
pytz
django-debug-toolbar
django-crispy-forms
django-widget-tweaks
django-model-utils
django-simple-history